api/serializers.py
from rest_framework import serializers
from contacts.models import Contact

class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = ['id', 'first_name', 'last_name', 'email', 'phone_number', 
                 'service', 'other_service', 'description', 'serving_status', 'created_at'] 