requirements.txt
Django
python-dotenv
pyngrok
django-cors-headers
djangorestframework 